+++ src/main/java/com/smunity/api/domain/petition/domain/Petition.java
@@ -1,6 +1,7 @@
 package com.smunity.api.domain.petition.domain;
 
 import com.smunity.api.domain.account.domain.User;
+import com.smunity.api.global.common.BaseEntity;
 import lombok.*;
 
 import javax.persistence.*;
@@ -11,7 +12,7 @@ import java.time.LocalDateTime;
 @Data
 @NoArgsConstructor
 @Table(name = "petitions_petition")
-public class Petition extends BaseEntity{
+public class Petition extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
